Enhance shenyu-plugin-grpc module unit test.
