+++ shenyu-plugin/shenyu-plugin-grpc/src/test/java/org/apache/shenyu/plugin/grpc/proto/MyStreamObserver.java
@@ -1,0 +1,56 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.shenyu.plugin.grpc.proto;
+
+import io.grpc.stub.StreamObserver;
+import lombok.SneakyThrows;
+
+public class MyStreamObserver implements StreamObserver<Boolean> {
+
+    private Boolean state;
+
+    public MyStreamObserver(Boolean state) {
+        this.state = state;
+    }
+
+    @SneakyThrows
+    @Override
+    public void onNext(Boolean value) {
+        if (!value) {
+            throw new Exception("exception");
+        } else {
+            state = true;
+        }
+    }
+
+    @SneakyThrows
+    @Override
+    public void onError(Throwable t) {
+        throw new Exception("exception");
+    }
+
+    @SneakyThrows
+    @Override
+    public void onCompleted() {
+        throw new Exception("exception");
+    }
+
+    public Boolean getState() {
+        return state;
+    }
+}
